git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene4765@1446985 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/queries/src/java/org/apache/lucene/queries/function/valuesource/IntFieldSource.java
@@ -85,7 +85,7 @@ public class IntFieldSource extends FieldCacheSource {
 
       @Override
       public String strVal(int doc) {
-        return Float.toString(arr.get(doc));
+        return Integer.toString(arr.get(doc));
       }
 
       @Override
